+++ Runtime/CoroutineExtentions.cs
@@ -10,168 +10,176 @@ namespace Jeomseon.Coroutine
 
     public static class CoroutineExtensions
     {
-        #region 공개 인터페이스
-        public static void StopCoroutineIfNotNull(this MonoBehaviour monoBehaviour, Coroutine coroutine)
+        public static void StopIfRunning(this MonoBehaviour host, Coroutine coroutine)
         {
-            if (!monoBehaviour || coroutine is null) return;
+            if (!host || coroutine is null) return;
 
-            monoBehaviour.StopCoroutine(coroutine);
+            host.StopCoroutine(coroutine);
         }
 
-        public static void StopCoroutineIfNotNull(this MonoBehaviour monoBehaviour, IEnumerator enumerator)
+        public static void StopIfRunning(this MonoBehaviour host, IEnumerator routine)
         {
-            if (!monoBehaviour || enumerator is null) return;
-            
-            monoBehaviour.StopCoroutine(enumerator);
+            if (!host || routine is null) return;
+
+            host.StopCoroutine(routine);
         }
 
-        public static Coroutine DoCallWaitForOneFrame(this MonoBehaviour monoBehaviour, Action action)
-         => monoBehaviour.StartCoroutine(iEDoCallWaitForOneFrame(action));
+        public static Coroutine InvokeNextFrame(this MonoBehaviour host, Action callback)
+            => host.StartCoroutine(InvokeNextFrameRoutine(callback));
 
-        public static Coroutine DoCallWaitForSeconds(this MonoBehaviour monoBehaviour, float delayTime, Action action)
-            => monoBehaviour.StartCoroutine(iEDoCallWaitForSeconds(delayTime, action));
+        public static Coroutine InvokeAfterSeconds(this MonoBehaviour host, float delay, Action callback)
+            => host.StartCoroutine(InvokeAfterSecondsRoutine(delay, callback));
 
-        public static Coroutine DoCallRoofCoroutine<T>(this MonoBehaviour monoBehaviour, Func<bool> action) where T : YieldInstruction
-            => monoBehaviour.StartCoroutine(iEDoCallRoofCoroutine<T>(action));
+        public static Coroutine RepeatWhile<TYieldInstruction>(this MonoBehaviour host, Func<bool> continueCondition)
+            where TYieldInstruction : YieldInstruction
+            => host.StartCoroutine(RepeatWhileRoutine<TYieldInstruction>(continueCondition));
 
-        public static Coroutine DoCallRoofCoroutine<T>(this MonoBehaviour monoBehaviour, Func<bool> action, Action init) where T : YieldInstruction
-            => monoBehaviour.StartCoroutine(iEDoCallRoofCoroutine<T>(action, init));
+        public static Coroutine RepeatWhile<TYieldInstruction>(this MonoBehaviour host, Func<bool> continueCondition, Action onStarted)
+            where TYieldInstruction : YieldInstruction
+            => host.StartCoroutine(RepeatWhileRoutine<TYieldInstruction>(continueCondition, onStarted));
 
-        public static Coroutine DoCallRoofCoroutine<T>(this MonoBehaviour monoBehaviour, Func<bool> action, Action init, Action finish) where T : YieldInstruction
-            => monoBehaviour.StartCoroutine(iEDoCallRoofCoroutine<T>(action, init, finish));
+        public static Coroutine RepeatWhile<TYieldInstruction>(this MonoBehaviour host, Func<bool> continueCondition, Action onStarted, Action onFinished)
+            where TYieldInstruction : YieldInstruction
+            => host.StartCoroutine(RepeatWhileRoutine<TYieldInstruction>(continueCondition, onStarted, onFinished));
 
-        public static Coroutine DoCallRoofCoroutineFinish<T>(this MonoBehaviour monoBehaviour, Func<bool> action, Action finish) where T : YieldInstruction
-            => monoBehaviour.StartCoroutine(iEDoCallRoofCoroutineFinish<T>(action, finish));
+        public static Coroutine RepeatWhileWithCompletion<TYieldInstruction>(this MonoBehaviour host, Func<bool> continueCondition, Action onFinished)
+            where TYieldInstruction : YieldInstruction
+            => host.StartCoroutine(RepeatWhileRoutine<TYieldInstruction>(continueCondition, null, onFinished));
 
-        public static Coroutine WaitCompletedConditions(this MonoBehaviour monoBehaviour, Func<bool> match, Action callback)
-            => monoBehaviour.StartCoroutine(iECallWaitCompletedConditions(match, callback));
+        public static Coroutine InvokeWhen(this MonoBehaviour host, Func<bool> predicate, Action callback)
+            => host.StartCoroutine(InvokeWhenRoutine(predicate, callback));
 
-        public static Coroutine WaitCompletedAsync(this MonoBehaviour monoBehaviour, Action asyncAction, Action callback)
-            => monoBehaviour.StartCoroutine(iEWaitCompletedAsyncTask(Task.Run(asyncAction), callback));
+        public static Coroutine RunInBackground(this MonoBehaviour host, Action backgroundWork, Action callback)
+            => host.StartCoroutine(WaitForTaskRoutine(Task.Run(backgroundWork), callback));
 
-        public static Coroutine WaitCompletedAsync<T>(this MonoBehaviour monoBehaviour, Func<T> asycnAction, Action<T> callback)
-            => monoBehaviour.StartCoroutine(iEWaitCompletedAsyncTask(asycnAction, callback));
+        public static Coroutine RunInBackground<T>(this MonoBehaviour host, Func<T> backgroundWork, Action<T> callback)
+            => host.StartCoroutine(WaitForTaskRoutine(Task.Run(backgroundWork), callback));
 
-        public static Coroutine ProgressFromEnumerable<T>(this MonoBehaviour monoBehaviour, IEnumerable<T> objects, Action<T> callback)
-            => monoBehaviour.StartCoroutine(iEProgressFromEnumerable(objects, callback));
+        public static Coroutine ProcessEachFrame<T>(this MonoBehaviour host, IEnumerable<T> items, Action<T> callback)
+            => host.StartCoroutine(ProcessEachFrameRoutine(items, callback));
 
-        public static Coroutine GetWaitComponent<T>(this MonoBehaviour monoBehaviour, Action<T> foundAction) where T : Component
-            => monoBehaviour.StartCoroutine(iEGetComponent(monoBehaviour, foundAction));
+        public static Coroutine GetComponentWithTimeout<T>(this MonoBehaviour host, Action<T> callback) where T : Component
+            => host.StartCoroutine(GetComponentWithTimeoutRoutine(host, 2.0f, callback));
 
-        public static Coroutine GetWaitComponent<T>(this MonoBehaviour monoBehaviour, float delayTime, Action<T> foundAction) where T : Component
-            => monoBehaviour.StartCoroutine(iEGetComponent(monoBehaviour, delayTime, foundAction));
+        public static Coroutine GetComponentWithTimeout<T>(this MonoBehaviour host, float timeout, Action<T> callback) where T : Component
+            => host.StartCoroutine(GetComponentWithTimeoutRoutine(host, timeout, callback));
 
-        public static IEnumerator AddCallback(this Coroutine coroutine, Action callback)
+        public static IEnumerator InvokeAfter(this Coroutine coroutine, Action callback)
         {
             yield return coroutine;
             callback.Invoke();
         }
 
-        public static IEnumerator AddCallback(this IEnumerator enumerator, Action callback)
+        public static IEnumerator InvokeAfter(this IEnumerator routine, Action callback)
         {
-            yield return enumerator;
-            callback.Invoke();
-        }
-#endregion
-
-        private static IEnumerator iECallWaitCompletedConditions(Func<bool> match, Action callback)
-        {
-            yield return new WaitUntil(match);
+            yield return routine;
             callback.Invoke();
         }
 
-        private static IEnumerator iEWaitCompletedAsyncTask(Task task, Action callback)
+        private static IEnumerator InvokeWhenRoutine(Func<bool> predicate, Action callback)
         {
-            yield return task;
+            yield return new WaitUntil(predicate);
             callback.Invoke();
         }
 
-        private static IEnumerator iEWaitCompletedAsyncTask<T>(Func<T> asyncAction, Action<T> callback)
+        private static IEnumerator WaitForTaskRoutine(Task task, Action callback)
         {
-            T someObject = default(T);
+            while (!task.IsCompleted)
+            {
+                yield return null;
+            }
 
-            yield return Task.Run(() => someObject = asyncAction.Invoke());
-            callback.Invoke(someObject);
+            if (task.IsFaulted)
+            {
+                throw task.Exception;
+            }
+
+            if (!task.IsCanceled)
+            {
+                callback.Invoke();
+            }
         }
 
-        private static IEnumerator iEProgressFromEnumerable<T>(IEnumerable<T> objects, Action<T> callback)
+        private static IEnumerator WaitForTaskRoutine<T>(Task<T> task, Action<T> callback)
         {
-            foreach (T item in objects)
+            while (!task.IsCompleted)
+            {
+                yield return null;
+            }
+
+            if (task.IsFaulted)
+            {
+                throw task.Exception;
+            }
+
+            if (!task.IsCanceled)
+            {
+                callback.Invoke(task.Result);
+            }
+        }
+
+        private static IEnumerator ProcessEachFrameRoutine<T>(IEnumerable<T> items, Action<T> callback)
+        {
+            foreach (T item in items)
             {
                 callback.Invoke(item);
                 yield return null;
             }
         }
 
-        private static IEnumerator iEDoCallWaitForOneFrame(Action action)
+        private static IEnumerator InvokeNextFrameRoutine(Action callback)
         {
             yield return null;
-            action.Invoke();
+            callback.Invoke();
         }
 
-        private static IEnumerator iEDoCallWaitForSeconds(float delayTime, Action action)
+        private static IEnumerator InvokeAfterSecondsRoutine(float delay, Action callback)
         {
-            yield return CoroutineHelper.WaitForSeconds(delayTime);
-            action.Invoke();
+            yield return CoroutineHelper.WaitForSeconds(delay);
+            callback.Invoke();
         }
 
-        private static IEnumerator iEDoCallRoofCoroutine<T>(Func<bool> action) where T : YieldInstruction
+        private static IEnumerator RepeatWhileRoutine<TYieldInstruction>(Func<bool> continueCondition)
+            where TYieldInstruction : YieldInstruction
         {
-            T yieldInstruction = getYieldInstruction<T>();
+            TYieldInstruction yieldInstruction = GetReusableYieldInstruction<TYieldInstruction>();
 
             do
             {
                 yield return yieldInstruction;
-            } while (action.Invoke());
+            } while (continueCondition.Invoke());
         }
 
-        private static IEnumerator iEDoCallRoofCoroutine<T>(Func<bool> action, Action init) where T : YieldInstruction
+        private static IEnumerator RepeatWhileRoutine<TYieldInstruction>(Func<bool> continueCondition, Action onStarted)
+            where TYieldInstruction : YieldInstruction
         {
-            init.Invoke();
-            yield return iEDoCallRoofCoroutine<T>(action);
+            onStarted?.Invoke();
+            yield return RepeatWhileRoutine<TYieldInstruction>(continueCondition);
         }
 
-        private static IEnumerator iEDoCallRoofCoroutine<T>(Func<bool> action, Action init, Action finish) where T : YieldInstruction
+        private static IEnumerator RepeatWhileRoutine<TYieldInstruction>(Func<bool> continueCondition, Action onStarted, Action onFinished)
+            where TYieldInstruction : YieldInstruction
         {
-            init.Invoke();
-            yield return iEDoCallRoofCoroutine<T>(action);
-            finish.Invoke();
+            onStarted?.Invoke();
+            yield return RepeatWhileRoutine<TYieldInstruction>(continueCondition);
+            onFinished?.Invoke();
         }
 
-        private static IEnumerator iEDoCallRoofCoroutineFinish<T>(Func<bool> action, Action finish) where T : YieldInstruction
-        {
-            yield return iEDoCallRoofCoroutine<T>(action);
-            finish.Invoke();
-        }
-
-        private static IEnumerator iEGetComponent<T>(MonoBehaviour monoBehaviour, Action<T> foundAction) where T : Component
-        {
-            T component;
-            float time = 2.0f;
-
-            while (!monoBehaviour.TryGetComponent(out component) && time > 0.0f)
-            {
-                yield return null;
-                time -= Time.deltaTime;
-            }
-
-            foundAction.Invoke(component);
-        }
-
-        private static IEnumerator iEGetComponent<T>(MonoBehaviour monoBehaviour, float delayTime, Action<T> foundAction) where T : Component
+        private static IEnumerator GetComponentWithTimeoutRoutine<T>(MonoBehaviour host, float timeout, Action<T> callback)
+            where T : Component
         {
             T component;
 
-            while (!monoBehaviour.TryGetComponent(out component) && delayTime > 0.0f)
+            while (!host.TryGetComponent(out component) && timeout > 0.0f)
             {
                 yield return null;
-                delayTime -= Time.deltaTime;
+                timeout -= Time.deltaTime;
             }
 
-            foundAction.Invoke(component);
+            callback.Invoke(component);
         }
 
-        private static T getYieldInstruction<T>() where T : YieldInstruction => typeof(T) switch
+        private static T GetReusableYieldInstruction<T>() where T : YieldInstruction => typeof(T) switch
         {
             { } type when type == typeof(WaitForEndOfFrame) => CoroutineHelper.WaitForEndOfFrame as T,
             { } type when type == typeof(WaitForFixedUpdate) => CoroutineHelper.WaitForFixedUpdate as T,